refactor: move AISettings to src/types/aisettings.ts
