+++ src/components/AISettings.tsx
@@ -13,27 +13,10 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { PromptEditor } from "./PromptEditor";
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger, } from "@/components/ui/accordion";
 import { configManager } from "@/utils/configManager";
-import { AppConfig, Endpoint, InferenceSettings } from "@/types/config";
+import { AppConfig, Endpoint } from "@/types/config";
+import { AISettings, DEFAULT_AI_SETTINGS } from "@/types/aisettings";
 
-// Re-export for compatibility with other components
-export interface AISettings {
-  apiKey: string;
-  apiUrl: string;
-  model: string;
-  provider: string;
-  inferenceSettings: InferenceSettings;  // Proper reference
-}
-
-export const DEFAULT_AI_SETTINGS: AISettings = {
-  apiKey: "",
-  apiUrl: "https://openrouter.ai/api",
-  model: "openrouter/free",
-  provider: "openrouter",
-  inferenceSettings: {  // ✅ Correct nesting
-    maxTokens: 800,
-    temp: 0.7            // ✅ Correct property name
-  }
-};
+export { AISettings, DEFAULT_AI_SETTINGS };
 
 interface AISettingsProps {
   onSettingsChange: (settings: AISettings) => void;
