+++ src/components/CharacterForm/PersonalitySection.tsx
@@ -6,7 +6,7 @@ import { Textarea } from "@/components/ui/textarea";
 import { Button } from "@/components/ui/button";
 import { Sparkles, Loader2, RefreshCcw, Trash2, X } from "lucide-react";
 import { generateWithAI, generatePersonality, generateScenario, generateFirstMes, generateMesExample } from "@/utils/aiGenerator";
-import { AISettings } from "@/components/AISettings";
+import { AISettings } from "@/types/aisettings";
 import { useToast } from "@/hooks/use-toast";
 
 interface PersonalitySectionProps {
