+++ src/pages/Index.tsx
@@ -6,7 +6,8 @@ import { useToast } from "@/hooks/use-toast";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { useLanguage } from "@/contexts/LanguageContext";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import AISettings, { AISettings as AISettingsType, DEFAULT_AI_SETTINGS } from "@/components/AISettings";
+import AISettings from "@/components/AISettings";
+import { AISettings as AISettingsType, DEFAULT_AI_SETTINGS } from "@/types/aisettings";
 import Toolbar from "@/components/Toolbar";
 import BasicInfoSection from "@/components/CharacterForm/BasicInfoSection";
 import PersonalitySection from "@/components/CharacterForm/PersonalitySection";
