+++ src/types/aisettings.ts
@@ -1,0 +1,21 @@
+import { InferenceSettings } from "@/types/config";
+
+// Re-export for compatibility with other components
+export interface AISettings {
+  apiKey: string;
+  apiUrl: string;
+  model: string;
+  provider: string;
+  inferenceSettings: InferenceSettings;  // Proper reference
+}
+
+export const DEFAULT_AI_SETTINGS: AISettings = {
+  apiKey: "",
+  apiUrl: "https://openrouter.ai/api",
+  model: "openrouter/free",
+  provider: "openrouter",
+  inferenceSettings: {  // ✅ Correct nesting
+    maxTokens: 800,
+    temp: 0.7            // ✅ Correct property name
+  }
+};
